refactor: make a function for formatting request body for a single endpoint
